add sidebar
