+++ app/(pages)/dashboard/edit/page.tsx
@@ -1,0 +1,13 @@
+import Dashboard from "../page";
+
+const EditTenantPage = () => {
+  return (
+    <Dashboard>
+      <div className="w-full overflow-scroll">
+        テナント編集ページ{/* <EditTenantForm /> */}
+      </div>
+    </Dashboard>
+  );
+};
+
+export default EditTenantPage;
